extract common functionality of searchable selects into mixin
